backend/function_ingest-farm-data/requirements.txt
functions-framework==3.*
flask==3.*
google-cloud-storage
google-cloud-bigquery
